backend/requirements.txt
fastapi
uvicorn
python-multipart
langchain
langchain-community
langchain-huggingface
langchain-chroma
pymupdf
sentence-transformers
python-dotenv